requirements.txt
onnxruntime-gpu==1.22.0
pillow==11.2.1